SGPD fixes, refactorings, se incluye cobrador lov en ticket

+++ pagodiario-domain/src/main/java/com/icetea/manager/pagodiario/dao/BillDao.java
@@ -16,7 +16,7 @@ public interface BillDao extends BasicIdentificableDao<Bill> {
 
 	List<Bill> findByCollectorId(Long collectorId);
 
-	List<Bill> find(Integer collectorId, Date dateFrom, Date dateTo);
+	List<Bill> find(Long collectorId, Date dateFrom, Date dateTo);
 
 	List<Bill> findExpires();
